+++ openpecha/__init__.py
@@ -1,2 +1,2 @@
 # Version of openpecha-toolkit package
-__version__ = "0.6.23"
+__version__ = "0.6.25"